validate the token and refactoring the component
